Use git:// instead of tarball for GitHub locations
Squashed with frznlogic path fix

It turns out that GitHub tarballs that have not been explicitly
uploaded but are simply provided by GitHub automatically for every tag,
have been confirmed to change MD5 sum.

Please refer to commit sent upstream for more details.

Note: This commit (on a fork) should be left behind when upgrading to
the tip of the upstream master, since they have merged the PR and
the PR modifies all the recipes and updates OTP*.

*In contrast, this commit is holding back the OTP version to 18.2.3
which is the one we had before the checksum issue.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>